# Split node min range is not stringent.