@yongfeng-nv 2020-02-17
Regarding to the failure (https://ci.tvm.ai/blue/organizations/jenkins/tvm/detail/PR-4885/1/pipeline), the underline issue is irrelevant.  This PR happens to exposed the bug, when it removes "if" statements.  I have created an issue (https://github.com/apache/incubator-tvm/issues/4890).  In such situation, how to proceed?