@tqchen 2020-02-19
In this case, unfortunately we need to wait until #4890 is fixed