@yongfeng-nv 2020-02-25
@Hzfengsy, I am going to close this PR without merging it, because it revealed more failures after our discussion.  Those failures require changing existing TOPI schedules because of thread binding after split.  I made another attempt (https://github.com/apache/incubator-tvm/pull/4931) with more sophisticated predicate than this one before tightening the extent.  

I have checked most of the cases we discussed here with the new change.  Could you review it (https://github.com/apache/incubator-tvm/pull/4931)?  Thank you.